added an additional button to id check based on feedback

+++ views/buttons/idverifybutton.py
@@ -9,7 +9,7 @@ class IdVerifyButton(discord.ui.View) :
 	def __init__(self) :
 		super().__init__(timeout=None)
 
-	@discord.ui.button(label="ID Verify", style=discord.ButtonStyle.green, custom_id="id_verify")
+	@discord.ui.button(label="Confirm ID Verify", style=discord.ButtonStyle.green, custom_id="id_verify")
 	async def idverify(self, interaction: discord.Interaction, button: discord.ui.Button) :
 		if not interaction.user.guild_permissions.administrator :
 			return await send_response(interaction, "You must have the administrator permission to execute this action!", ephemeral=True)
@@ -25,4 +25,10 @@ class IdVerifyButton(discord.ui.View) :
 			await send_response(interaction,f"Can't find entry: <@{user.id}>")
 			return
 		await send_response(interaction, f"Deleted entry: <@{user.id}>")
+		await interaction.message.delete()
+
+	@discord.ui.button(label="User Left", style=discord.ButtonStyle.red, custom_id="left")
+	async def remove(self, interaction: discord.Interaction, button: discord.ui.Button) :
+		if not interaction.user.guild_permissions.administrator :
+			return await send_response(interaction, "You must have the administrator permission to execute this action!", ephemeral=True)
 		await interaction.message.delete()
